requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2

# Markdown
markdown==3.5.1
Pygments==2.17.2

# Templates
Jinja2==3.1.2

# YAML
PyYAML==6.0.1

# Configuration
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
